Add GitHub Actions workflow and ARM templates for API
Added a new GitHub Actions workflow configuration file
(BookAPI20241006223516.yml) to build and deploy a .NET Core
application to an Azure Web App and import the API into Azure
API Management Service. The workflow triggers on pushes to the
main branch and consists of two jobs: build and deploy.

Added service dependencies configuration files
(serviceDependencies.BookAPI20241006223516.json and
serviceDependencies.json) to define dependencies for the API
Management service.

Added a new ARM template file (apis1.arm.json) to define the
infrastructure for the API Management service and the API
itself, including parameters for resource group name and
location, and configurations for the API Management service
and the API.