Add JSON config and ARM template for BeverageAPI2024
Introduced a new JSON configuration for `serviceDependencies.BeverageAPI2024 - Web Deploy.json`, specifying dependencies for the BeverageAPI2024 service, including an Azure API service (`apis1`), an identity application (`identityapp1`), and an MSSQL database (`mssql1`) with a connection string identifier.

Added a new ARM template in `profile.arm.json` to define the deployment of an Azure App Service for the BeverageAPI2024 application. This template includes parameters for resource group name and location, resource name, and resource location. It also defines variables for the App Service Plan name and resource ID, and specifies the creation of a resource group and an incremental deployment of the App Service, including its configuration, identity, and associated App Service Plan.