Add Role column to Users; remove sensitive logging
Added a nullable integer column `Role` with a default value of 3 to the `Users` table in the database schema. This change is reflected in `BeverageContextModelSnapshot.cs` and new migration files `20241021144323_AddedRolesUser.cs` and `20241021144323_AddedRolesUser.Designer.cs`.

Removed the retrieval and logging of the JWT secret key from `Program.cs` to enhance security. Also removed the code that iterated over configuration sources and printed them to the console to prevent exposure of sensitive configuration data.